Update
Code cleanup for compilation with Clang and MSVC
Add user service layer
Add an option to disable default key behavior
Disable notifier surface from rendering
Instanciable Tracer
Refactor Tracer initilization with logger to use std location
Update primary services identification
Improve Window fullscreen management
Adds C++ constants instead of macros
Add the visibility check per surfaces for input in overlay manager.
Enable Wayland support.
Update some constexpr usage.
GTK3 removed.
Add dialog box cancelable state.
Crash fixed on some situation.
Shader generators flexibility improved.
Renderable instance code improved.
Semaphores and fences improved in render service.
Compatible with PCH compiler feature.

+++ src/DownloadItem.cpp
@@ -34,7 +34,9 @@ namespace EmEn
 	using namespace EmEn::Libs;
 
 	DownloadItem::DownloadItem (Network::URL url, std::filesystem::path output, bool replaceExistingFile) noexcept
-		: m_url(std::move(url)), m_output(std::move(output)), m_replaceExistingFile(replaceExistingFile)
+		: m_url(std::move(url)),
+		m_output(std::move(output)),
+		m_replaceExistingFile(replaceExistingFile)
 	{
 
 	}
@@ -58,7 +60,7 @@ namespace EmEn
 	}
 
 	void
-	DownloadItem::setProgression (size_t total, size_t received) noexcept
+	DownloadItem::setProgression (uint64_t total, uint64_t received) noexcept
 	{
 		m_bytesTotal = total;
 		m_bytesReceived = received;
